Added new implementations of Expression

+++ src/Expression.java
@@ -1,5 +1,5 @@
 import java.util.Map;
 
 public interface Expression {
-	public int interpret(Map<String, Expression> variables);
+	public double interpret(Map<String, Expression> variables);
 }
